refactor(hub): deepen HTTP client module
